requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
bson==0.5.10
cached-property==1.5.2
celery==5.1.1
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.9
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.4
django-common-helpers==0.9.2
django-cron==0.5.1
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.13.1
exponent-server-sdk==2.0.0
future==0.18.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
importlib-metadata==4.5.0
kombu==5.1.0
marshmallow==3.14.1
motor==2.4.0
numpy==1.21.4
packaging==21.3
pandas==1.1.5
promise==2.3
prompt-toolkit==3.0.19
psycopg2==2.9.2
pydantic==1.7.4
pymongo==3.11.4
pyparsing==3.0.7
python-dateutil==2.8.1
python-decouple==3.5
python-dotenv==0.17.1
pytz==2021.3
redis==4.1.2
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
starlette==0.13.6
typing-extensions==3.10.0.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.13.4
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wincertstore==0.2
wrapt==1.13.3
zipp==3.4.1
django-cors-headers==3.11.0